// src/types/Experience.ts
export interface Experience {
  id: string;
  company: string;
  position: string;
  period: string;
  description: string;
  technologies: string[];
}

export interface ExperienceCardType {
  name: string;
  data: string;
  company: string;
  description: string;
  skills: string[];
  color: string;
}